Reorganized components into individual folders
